create middleware file to make things clean
