+++ server.js
@@ -4,6 +4,7 @@ var http = require('http')
   , path = require('path')
   , logger = require('./logger').get('appium')
   , appium = require('./app/appium')
+  , bodyParser = require('./middleware').parserWrap
   , parser = require('./app/parser');
 
 var main = function(args, readyCb, doneCb) {
@@ -16,30 +17,12 @@ var main = function(args, readyCb, doneCb) {
   args.device = 'iOS';
 
   rest.configure(function() {
-    var bodyParser = express.bodyParser()
-      , parserWrap = function(req, res, next) {
-          // wd.js sends us http POSTs with empty body which will make bodyParser fail.
-          var cLen = req.get('content-length');
-          if (typeof cLen === "undefined" || parseInt(cLen, 10) <= 0) {
-            req.headers['content-length'] = 0;
-            next();
-          } else {
-            // hack because python client library sux
-            if (req.headers['content-type'] === 'application/x-www-form-urlencoded') {
-              req.headers['content-type'] = 'application/json';
-            }
-            bodyParser(req, res, next);
-          }
-        };
-    rest.use(function(req, res, next) {
-      next();
-    });
     rest.use(express.favicon());
     rest.use(express.static(path.join(__dirname, '/app/static')));
     if (args.verbose) {
       rest.use(express.logger('dev'));
     }
-    rest.use(parserWrap);
+    rest.use(bodyParser);
     rest.use(express.methodOverride());
     rest.use(rest.router);
   });
